Entrega Final

+++ Js/carrito.js
@@ -3,7 +3,7 @@ productosEnCarrito = JSON.parse(productosEnCarrito);
 
 const contenedorCarroVacio = document.querySelector('#carro-vacio');
 const contenedorCarroProductos = document.querySelector('#carro-productos');
-const contenedorCarroAcciones = document.querySelector('#carro-acciones');
+const contenedorCarroAcciones = document.querySelector('#carrito-acciones');
 let botonesEliminar = document.querySelectorAll('.carro-producto-eliminar');
 const botonVaciar = document.querySelector('#carrito-acciones-vaciar')
 const contenedorTotal = document.querySelector('#total')
@@ -13,18 +13,18 @@ function cargarProductosCarrito(){
 
     if (productosEnCarrito && productosEnCarrito.length > 0) {
 
-        contenedorCarroVacio.classList.add('disabled')
-        contenedorCarroProductos.classList.remove('disabled')
-        contenedorCarroAcciones.classList.remove('disabled')
+        contenedorCarroVacio.classList.add('disabled');
+        contenedorCarroProductos.classList.remove('disabled');
+        contenedorCarroAcciones.classList.remove('enabled');
         
-        contenedorCarroProductos.innerHTML = '';
+        contenedorCarroProductos.innerHTML = "";
     
         productosEnCarrito.forEach(producto => {
     
             const div = document.createElement('div');
             div.classList.add('carro-producto')
             div.innerHTML = `
-            <img class="carro-producto-imagen" src="${producto.imagen}" alt="${producto.titulo}">
+                            <img class="carro-producto-imagen" src="${producto.imagen}" alt="${producto.titulo}">
                             <div class="carro-titulo">
                                 <small>titulo</small>
                                 <h3>${producto.titulo}</h3>
@@ -43,9 +43,14 @@ function cargarProductosCarrito(){
                             </div>
                             <button class="carro-producto-eliminar" id=${producto.id}><i class="bi bi-trash"></i></button>
             `;
+
             contenedorCarroProductos.append(div);
+
         })
-    
+
+    actualizarBotonesEliminar();
+    actualizarTotal();
+
     } else {
     
         contenedorCarroVacio.classList.remove('disabled')
@@ -67,12 +72,36 @@ function actualizarBotonesEliminar() {
     });
 }
 
-function eliminarDelCarro(){
+function eliminarDelCarrito(e){
+
+    Toastify({
+        text: "Producto eliminado",
+        duration: 3000,
+        close: true,
+        gravity: "top",
+        position: "right",
+        stopOnFocus: true,
+        style: {
+        background: "linear-gradient(to right, #4b33a8, #785ce9)",
+        borderRadius: "2rem",
+        textTransform: "uppercase",
+        fontSize: ".75rem"
+        },
+        offset: {
+            x: '1.5rem',
+            y: '1.5rem'
+        },
+        onClick: function(){} 
+    }).showToast();
+
+
+
+
     const idBoton = e.currentTarget.id;
     const productoEliminado = productosEnCarrito.find(producto => producto.id === idBoton);
-    const index = productosEnCarrito.findeIndex(producto = producto.id === idBoton);
-    productosEnCarrito.splice(index, 1);
+    const index = productosEnCarrito.findIndex(producto => producto.id === idBoton);
 
+    productosEnCarrito.splice(index, 1);
     cargarProductosCarrito();
 
     localStorage.setItem('productos-en-carrito', JSON.stringify(productosEnCarrito)) ;   
@@ -80,10 +109,24 @@ function eliminarDelCarro(){
 
 botonVaciar.addEventListener('click', vaciarCarrito);
 function vaciarCarrito() {
-    productosEnCarrito.length = 0;
-    localStorage.setItem('productos-en-carrito', JSON.stringify(productosEnCarrito));
-    cargarProductosCarrito();
+
+    Swal.fire({
+        title: '¿Estás seguro?',
+        icon: 'question',
+        html: `Se van a borrar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+        showCancelButton: true,
+        focusConfirm: false,
+        confirmButtonText: 'Sí',
+        cancelButtonText: 'No'
+    }).then((result) => {
+        if (result.isConfirmed) {
+            productosEnCarrito.length = 0;
+            localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+            cargarProductosCarrito();
+        }
+    })
 }
+
 
 function actualizarTotal() {
     const totalCalculado =  productosEnCarrito.reduce ((acc,producto) => acc + (producto.precio * producto.cantidad), 0);
@@ -91,7 +134,7 @@ function actualizarTotal() {
 }
 
 botonComprar.addEventListener('click', comprarCarrito);
-function vaciarCarrito() {
+function comprarCarrito() {
 
     productosEnCarrito.length = 0;
     localStorage.setItem('productos-en-carrito', JSON.stringify(productosEnCarrito));
